requirements.txt
Babel==2.5.3
Django==2.0.3
django-phonenumber-field==2.0.0
phonenumberslite==8.9.2
pytz==2018.3
